Code cleaning and initial commit
